requirements.txt
streamlit
transformers
deepface
emoji
Pillow
